Added hash join

+++ src/main/java/ru/bmstu/sqlfornosql/model/Table.java
@@ -152,7 +152,7 @@ public class Table {
         this.rows.clear();
     }
 
-    public long size() {
+    public int size() {
         return rows.size();
     }
